Add PubNubManager
